# docs/swagger.yaml
basePath: /
definitions:
  internal_auth.ErrorResponse:
    properties:
      error:
        properties:
          code:
            type: string
          message:
            type: string
        type: object
    type: object
  internal_auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  internal_auth.LoginResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      token_type:
        type: string
    type: object
  internal_auth.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  internal_auth.RegisterResponse:
    properties:
      data:
        properties:
          email:
            type: string
          user_id:
            type: string
        type: object
    type: object
...
  internal_organization.CreateRequest:
    properties:
      name:
        example: My Company
        maxLength: 255
        minLength: 1
        type: string
    required:
    - name
    type: object
  internal_organization.CreateResponse:
    properties:
      data:
        properties:
          avatar_url:
            type: string
          created_at:
            type: string
          id:
            type: string
          name:
            type: string
          owner_id:
            type: string
          slug:
            type: string
          status:
            type: string
          updated_at:
            type: string
        type: object
    type: object
  internal_organization.ErrorResponse:
    properties:
      error:
        properties:
          code:
            type: string
          message:
            type: string
        type: object
    type: object
  internal_organization.ListResponse:
    properties:
      data:
        properties:
          organizations:
            items:
              properties:
                avatar_url:
                  type: string
                created_at:
                  type: string
                id:
                  type: string
                name:
                  type: string
                owner_id:
                  type: string
                slug:
                  type: string
                status:
                  type: string
                updated_at:
                  type: string
              type: object
            type: array
        type: object
    type: object
...
  internal_team.CreateRequest:
    properties:
      description:
        example: Команда разработки бэкенда
        type: string
      name:
        example: Backend Team
        maxLength: 255
        minLength: 1
        type: string
      organization_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    required:
    - name
    - organization_id
    type: object
  internal_team.ErrorResponse:
    properties:
      error:
        properties:
          code:
            type: string
          message:
            type: string
        type: object
    type: object
  internal_team.TeamResponse:
    properties:
      data:
        properties:
          created_at:
            type: string
          creator_id:
            type: string
          description:
            type: string
          id:
            type: string
          name:
            type: string
          organization_id:
            type: string
        type: object
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Backend API для работы с командой над проектом.
  title: timeManager API
  version: "1.0"
paths:
  /api/v1/auth/{provider}/callback:
    get:
      description: Обрабатывает callback от OAuth-провайдера, создаёт/находит пользователя,
        редирект на фронтенд с токеном.
      parameters:
      - description: google | yandex | github
        in: path
        name: provider
        required: true
        type: string
      - description: Authorization code from provider
        in: query
        name: code
        required: true
        type: string
      - description: State (CSRF protection)
        in: query
        name: state
        required: true
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Redirect to frontend with ?token=...
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_auth.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_auth.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_auth.ErrorResponse'
      summary: OAuth — callback от провайдера
      tags:
      - auth
  /api/v1/auth/{provider}/redirect:
    get:
      description: 'Редирект на страницу авторизации Google, Yandex или GitHub. Поддерживаемые
        provider: google, yandex, github.'
      parameters:
      - description: google | yandex | github
        in: path
        name: provider
        required: true
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Redirect to OAuth provider
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_auth.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_auth.ErrorResponse'
      summary: OAuth — редирект на провайдера
      tags:
      - auth
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Проверяет учётные данные и возвращает JWT access token.
      parameters:
      - description: Email и пароль
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_auth.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_auth.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_auth.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_auth.ErrorResponse'
      summary: Вход в систему
      tags:
      - auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Создаёт нового пользователя по email и паролю. Пароль должен быть
        не менее 8 символов.
      parameters:
      - description: Email и пароль
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_auth.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_auth.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_auth.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_auth.ErrorResponse'
      summary: Регистрация пользователя
      tags:
      - auth
  /api/v1/files/{path}:
    get:
      description: 'Возвращает файл (картинку) из MinIO по частичному пути. Путь:
        users/{id}/avatar.jpg, organizations/{id}/avatar.jpg, teams/{id}/avatar.jpg'
      parameters:
      - description: Частичный путь к файлу (например users/uuid/avatar.jpg)
        in: path
        name: path
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Файл
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Получить файл из хранилища
      tags:
      - files
  /api/v1/organizations:
    get:
      description: Возвращает организации, в которых текущий пользователь является
        участником.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_organization.ListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_organization.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_organization.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Список моих организаций
      tags:
      - organizations
    post:
      consumes:
      - application/json
      - multipart/form-data
      description: Создаёт организацию от имени текущего пользователя. Slug генерируется
        автоматически из name. Поддерживает JSON или multipart/form-data (name + опционально
        avatar).
      parameters:
      - description: 'Название организации (JSON) или multipart: name, avatar'
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_organization.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_organization.CreateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_organization.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_organization.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_organization.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создать организацию
      tags:
      - organizations
...
  /api/v1/teams:
    post:
      consumes:
      - application/json
      - multipart/form-data
      description: Создаёт команду в организации. Создатель должен быть членом организации.
        Поддерживает JSON или multipart (name, description, organization_id + опционально
        avatar).
      parameters:
      - description: Данные команды
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_team.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_team.TeamResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_team.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_team.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_team.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_team.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_team.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создать команду
      tags:
      - teams
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"